Fat refactor of Tokens section (#2554)
- breaks up tokens section into multiple components
- makes project and projects page more html-semantic
- small style fixes and improvements along the way
