+++ src/components/v2v3/V2V3Project/ProjectActivity/index.tsx
@@ -354,7 +354,7 @@ export default function ProjectActivity() {
   }, [isLoading, isFetchingNextPage, hasNextPage, fetchNextPage, colors, count])
 
   return (
-    <div style={{ marginTop: '20px' }}>
+    <div>
       <div
         style={{
           display: 'flex',
