+++ src/components/v2v3/V2V3Project/V2V3FundingCycleSection/PayoutSplitsCard.tsx
@@ -151,9 +151,9 @@ export default function PayoutSplitsCard({
           >
             <TooltipLabel
               label={
-                <h4 style={{ display: 'inline-block' }}>
+                <h3 style={{ display: 'inline-block', fontSize: '0.875rem' }}>
                   <Trans>Funding distribution</Trans>
-                </h4>
+                </h3>
               }
               tip={
                 <Trans>
