+++ src/components/v2v3/V2V3Project/V2V3ManageTokensSection/V1ProjectTokensDescription/index.tsx
@@ -1,0 +1,1 @@
+export * from './V1ProjectTokensDescription'
